requirements.txt
Flask==3.1.0
requests==2.32.3
gunicorn==23.0.0
python-dotenv==1.0.0
imgkit==1.2.3

#other dependencies sudo apt-get install wkhtmltopdf
#pip install flask requests imgkit python-dotenv
